Split bindings into two mixins
