+++ src/can-stache-element.js
@@ -5,6 +5,7 @@ const mixinProps = require("./mixin-props");
 const mixinStacheView = require("./mixin-stache-view");
 const mixinViewModelSymbol = require("./mixin-viewmodel-symbol");
 const mixinBindings = require("./mixin-bindings");
+const mixinInitializeBindings = require("./mixin-initialize-bindings");
 
 const canStacheBindings = require("can-stache-bindings");
 
@@ -20,12 +21,15 @@ function DeriveElement(BaseElement = HTMLElement) {
 	mixinLifecycleMethods(
 		// mixin .bindings() method and behavior
 		mixinBindings(
-			// mix in viewModel symbol used by can-stache-bindings
-			mixinViewModelSymbol(
-				// mix in stache renderer from `static view` property
-				mixinStacheView(
-					// add getters/setters from `static props` property
-					mixinProps(BaseElement)
+			// Initialize the bindings
+			mixinInitializeBindings(
+				// mix in viewModel symbol used by can-stache-bindings
+				mixinViewModelSymbol(
+					// mix in stache renderer from `static view` property
+					mixinStacheView(
+						// add getters/setters from `static props` property
+						mixinProps(BaseElement)
+					)
 				)
 			)
 		)
