+++ src/mixin-bindings.js
@@ -1,13 +1,11 @@
 "use strict";
 
-const stacheBindings = require("can-stache-bindings");
 const keyObservable = require("can-simple-observable/key/key");
 const canReflect = require("can-reflect");
 const Bind = require("can-bind");
 
 const getValueSymbol = Symbol.for("can.getValue");
 const setValueSymbol = Symbol.for("can.setValue");
-const lifecycleStatusSymbol = Symbol.for("can.lifecycleStatus");
 const metaSymbol = Symbol.for("can.meta");
 
 module.exports = function mixinBindings(Base = HTMLElement) {
@@ -23,17 +21,19 @@ module.exports = function mixinBindings(Base = HTMLElement) {
 			var savedBindings = this[metaSymbol] && this[metaSymbol]._connectedBindings;
 			if (savedBindings) {
 				props = props || {};
-				const bindingContext = {
-					element: this
-				};
-				const bindings = [];
+
+				if (this[metaSymbol]._bindings === undefined) {
+					this[metaSymbol]._bindings = [];
+				}
 
 				canReflect.eachKey(savedBindings, (parent, propName) => {
+
 					var canGetParentValue = parent != null && !!parent[getValueSymbol];
 					var canSetParentValue = parent != null && !!parent[setValueSymbol];
 
 					// If we can get or set the value, then we’ll create a binding
 					if (canGetParentValue === true || canSetParentValue) {
+
 						// Create an observable for reading/writing the viewModel
 						// even though it doesn't exist yet.
 						var child = keyObservable(this, propName);
@@ -42,8 +42,7 @@ module.exports = function mixinBindings(Base = HTMLElement) {
 						var canBinding = new Bind({
 							child: child,
 							parent: parent,
-							queue: "dom",
-							element: this,
+							queue: "domUI",
 
 							//!steal-remove-start
 							// For debugging: the names that will be assigned to the updateChild
@@ -53,7 +52,7 @@ module.exports = function mixinBindings(Base = HTMLElement) {
 							//!steal-remove-end
 						});
 
-						bindings.push({
+						this[metaSymbol]._bindings.push({
 							binding: canBinding,
 							siblingBindingData: {
 								parent: {
@@ -68,51 +67,25 @@ module.exports = function mixinBindings(Base = HTMLElement) {
 								bindingAttributeName: propName
 							}
 						});
+
 					} else {
 						// Can’t get or set the value, so assume it’s not an observable
 						props[propName] = parent;
 					}
 				});
 
-				// Initialize the viewModel.  Make sure you
-				// save it so the observables can access it.
-				var initializeData = stacheBindings.behaviors.initializeViewModel(bindings, props, (properties) => {
-					super.initialize(properties);
-					return this;
-				}, bindingContext);
-
-				this[metaSymbol]._connectedBindingsTeardown = function() {
-					for (var attrName in initializeData.onTeardowns) {
-						initializeData.onTeardowns[attrName]();
-					}
-				};
-			} else {
-				if (super.initialize) {
-					super.initialize(props);
-				}
+				this[metaSymbol].other = true;
+			}
+			if (super.initialize) {
+				super.initialize(props);
 			}
 		}
-		render(props, renderOptions) {
+		render(props, renderOptions, parentNodeList) {
 			const viewRoot = this.viewRoot || this;
 			viewRoot.innerHTML = "";
 
 			if(super.render) {
-				super.render(props, renderOptions);
-			}
-		}
-		disconnect() {
-			if(this[metaSymbol] && this[metaSymbol]._connectedBindingsTeardown) {
-				this[metaSymbol]._connectedBindingsTeardown();
-				this[metaSymbol]._connectedBindingsTeardown = null;
-			}
-			this[lifecycleStatusSymbol] = {
-				initialized: false,
-				rendered: false,
-				connected: false,
-				disconnected: true
-			};
-			if (super.disconnect) {
-				super.disconnect();
+				super.render(props, renderOptions, parentNodeList);
 			}
 		}
 	};
